requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.1
djangorestframework==3.13.1
idna==3.4
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xmltodict==0.13.0
